feat: apply Canvas backgroundColor to clear color at runtime
Read backgroundColor from Canvas component each frame and call
setClearColor in SDK runtime, editor preview, and native C++ paths
instead of using a hardcoded value.

+++ src/esengine/ecs/components/Canvas.hpp
@@ -92,7 +92,7 @@ struct Canvas {
      * @return Half-height in world units for orthographic projection
      */
     f32 getOrthoSize() const {
-        return (designResolution.y * 0.5f) / pixelsPerUnit;
+        return (static_cast<f32>(designResolution.y) * 0.5f) / pixelsPerUnit;
     }
 
     /**
